+++ CHANGES.md
@@ -6,6 +6,9 @@ Fast and flexible PHP library rendering plain ASCII tables.
 
 ## Changes
 
+* @dev
+  * Corrected rendering tables with first/last column hidden.
+
 * 1.6.0 (2022-10-19)
   * Rows can now contain incomplete set of column cells.